renames config file
